move welcome screen into ui package solve gets its own package
