+++ src/main/java/com/canoo/cog/Main.java
@@ -20,14 +20,11 @@ package com.canoo.cog;
  * #L%
  */
 
-
 import java.io.IOException;
 
 import com.canoo.cog.sonar.SonarService;
+import com.canoo.cog.ui.welcome.WelcomeService;
 import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.Pane;
 import javafx.stage.Stage;
 
 public class Main extends Application {
@@ -38,22 +35,8 @@ public class Main extends Application {
 
     @Override
     public void start(Stage stage) throws IOException {
-
-        // Initialize Sonar Service
-        SonarService sonarService = new SonarService();
-
         // Initialize Welcome Screen
-        FXMLLoader loader = new FXMLLoader(getClass().getResource("welcome.fxml"));
-        WelcomeController welcomeController = new WelcomeController(sonarService);
-        loader.setController(welcomeController);
-        Pane welcomePane = loader.load();
-        WelcomeController controller = loader.getController();
-        controller.init();
-        Scene welcomeScene = new Scene(welcomePane);
-
-        // Initialize JavaFx Welcome Stage
-        stage.setTitle("Code of Gotham");
-        stage.setScene(welcomeScene);
-        stage.show();
+        final WelcomeService welcomeService = new WelcomeService(stage, new SonarService());
+        welcomeService.startScene();
     }
 }
