+++ src/main/resources/com/canoo/cog/ui/welcome/welcome.fxml
